+++ lib/src/attributes/widget_decorators/flexible.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mix/src/attributes/common/attribute.dart';
 import 'package:mix/src/mixer/mix_context.dart';
+
 /// ## Widget
 /// - (All)
 /// ## Utilities