+++ lib/src/headless/switch.dart
@@ -6,6 +6,7 @@ import 'package:mix/src/theme/tokens/color_scheme_tokens.dart';
 import 'package:mix/src/widgets/box.widget.dart';
 import 'package:mix/src/widgets/mixable.widget.dart';
 import 'package:mix/src/widgets/pressable.widget.dart';
+
 /// _Mix_ corollary to Flutter _Switch_ widget
 ///
 /// Default _Mix_ values: