+++ lib/src/helpers/extensions.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:mix/src/attributes/common/attribute.dart';
 import 'package:mix/src/attributes/text/text.attributes.dart';
 import 'package:mix/src/attributes/text/text.notifier.dart';
 import 'package:mix/src/dto/box_shadow.dto.dart';
@@ -92,5 +93,13 @@ extension Matrix4Extension on Matrix4 {
   Matrix4 merge(Matrix4? other) {
     if (other == null || other == this) return this;
     return clone()..multiply(other);
+  }
+}
+
+/// {@category Misc Utils}
+extension BoolExtension on bool {
+  Attribute? call(Attribute attribute) {
+    if (this) return attribute;
+    return null;
   }
 }