+++ lib/src/widgets/remixable.widget.dart
@@ -1,7 +1,0 @@
-import 'package:mix/mix.dart';
-
-class Remix<T extends RemixableWidget> {
-  const Remix._(this.widget);
-
-  final Type widget;
-}