+++ test/mixer/mix_context_test.dart
@@ -21,15 +21,13 @@ final pressableMix = Mix.fromList([
 
 class _MixContextTestWidget extends StatelessWidget {
   const _MixContextTestWidget({Key? key}) : super(key: key);
-  static bool hasBeenPressed = false;
+
   @override
   Widget build(BuildContext context) {
     return MixTestWidget(
       child: Pressable(
         mix: pressableMix,
-        onPressed: () {
-          hasBeenPressed = true;
-        },
+        onPressed: () {},
         child: Column(
           children: [
             const TextMix('Hello'),
@@ -42,7 +40,7 @@ class _MixContextTestWidget extends StatelessWidget {
               mix: Mix(
                 textColor(Colors.purple),
               ),
-            )
+            ),
           ],
         ),
       ),
@@ -67,11 +65,26 @@ void main() {
           mix: pressableMix,
         );
 
-        expect(mixContext.originalMix, pressableMix);
+        expect(mixContext.sourceMix, pressableMix);
+        expect(
+          mixContext.descendentMix,
+          matchContext.descendentMix,
+          reason: 'descendentMix',
+        );
 
-        expect(matchContext, mixContext);
-        expect(matchContext == mixContext, true);
-        // Different instance but same same properties
+        expect(
+          mixContext.originalMix,
+          matchContext.originalMix,
+          reason: 'originalMix',
+        );
+
+        expect(
+          mixContext.sourceMix,
+          matchContext.sourceMix,
+          reason: 'sourceMix',
+        );
+
+        // Different instance but same properties
         expect(matchContext.hashCode == mixContext.hashCode, false);
       });